# WideIO.API/WideIO.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "description" : "VideIO project API specification.",
    "title" : "VideIO API",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/zagorskim/VideIO/1.0.0"
  } ],
...
  "paths" : {
    "/user" : {
      "delete" : {
        "operationId" : "deleteUserData",
        "parameters" : [ {
          "description" : "User ID",
          "explode" : true,
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "summary" : "User account deletion",
        "tags" : [ "User" ]
      },
      "get" : {
        "operationId" : "getUserData",
        "parameters" : [ {
          "description" : "User ID",
          "explode" : true,
          "in" : "query",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "summary" : "User data retrieval",
        "tags" : [ "User" ]
      },
      "put" : {
        "operationId" : "editUserData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "summary" : "User data editing",
        "tags" : [ "User" ]
      }
    },
...
    "/register" : {
      "post" : {
        "operationId" : "registerUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          }
        },
        "summary" : "User registration",
        "tags" : [ "User" ]
      }
    },
    "/login" : {
      "post" : {
        "operationId" : "loginUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application:json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LoginResponseDto"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          }
        },
        "summary" : "User login",
        "tags" : [ "User" ]
      }
    },
    "/ban/{id}" : {
      "post" : {
        "operationId" : "banUser",
        "parameters" : [ {
          "description" : "User ID",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad request"
          },
          "401" : {
            "description" : "Unauthorized"
          }
        },
        "summary" : "Ban user",
        "tags" : [ "User" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserDto" : {
        "example" : {
          "surname" : "Doe",
          "nickname" : "johnny123",
          "name" : "John",
          "id" : "123e4567-e89b-12d3-a456-426614174000",
          "userType" : null,
          "accountBalance" : 0.8008281904610115,
          "email" : "john.doe@mail.com"
        },
        "properties" : {
          "id" : {
            "example" : "123e4567-e89b-12d3-a456-426614174000",
            "format" : "uuid",
            "type" : "string"
          },
          "email" : {
            "example" : "john.doe@mail.com",
            "type" : "string"
          },
          "nickname" : {
            "example" : "johnny123",
            "type" : "string"
          },
          "name" : {
            "example" : "John",
            "type" : "string"
          },
          "surname" : {
            "example" : "Doe",
            "type" : "string"
          },
          "accountBalance" : {
            "format" : "currency",
            "type" : "number"
          },
          "userType" : {
            "$ref" : "#/components/schemas/UserTypeDto"
          }
        },
        "type" : "object"
      },
      "RegisterDto" : {
        "example" : {
          "password" : "Passwd123!",
          "surname" : "Doe",
          "nickname" : "johnny123",
          "name" : "John",
          "email" : "john.doe@mail.com"
        },
        "properties" : {
          "email" : {
            "example" : "john.doe@mail.com",
            "type" : "string"
          },
          "nickname" : {
            "example" : "johnny123",
            "type" : "string"
          },
          "name" : {
            "example" : "John",
            "type" : "string"
          },
          "surname" : {
            "example" : "Doe",
            "type" : "string"
          },
          "password" : {
            "example" : "Passwd123!",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "LoginDto" : {
        "example" : {
          "password" : "Passwd123!",
          "email" : "john.doe@mail.com"
        },
        "properties" : {
          "email" : {
            "example" : "john.doe@mail.com",
            "type" : "string"
          },
          "password" : {
            "example" : "Passwd123!",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "LoginResponseDto" : {
        "properties" : {
          "token" : {
            "example" : "AbCdEf123456AbCdEf123456AbCdEf123456",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UserTypeDto" : {
        "enum" : [ "Simple", "Creator", "Administrator" ],
        "type" : "string"
      },
...
    }
  }
}